+++ codegen/masm/src/lib.rs
@@ -169,6 +169,7 @@ fn lower_hir_ops(info: &mut midenc_hir::DialectInfo) {
     info.register_operation_trait::<hir::ConstantPointer, dyn HirLowering>();
     info.register_operation_trait::<hir::Exec, dyn HirLowering>();
     info.register_operation_trait::<hir::ExecIndirect, dyn HirLowering>();
+    info.register_operation_trait::<hir::ProcedureRoot, dyn HirLowering>();
     info.register_operation_trait::<hir::ExecFpi, dyn HirLowering>();
     info.register_operation_trait::<hir::Call, dyn HirLowering>();
     info.register_operation_trait::<hir::Syscall, dyn HirLowering>();